% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_city}
\alias{china_city}
\title{city}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 476 rows and 10 columns.
}
\usage{
china_city
}
\description{
city data for china
source: http://datav.aliyun.com/tools/atlas/

\describe{
  \item{name}{city name}
  \item{adcode}{adcode}
  ...
}
}
\keyword{datasets}
