% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_province}
\alias{china_province}
\title{province}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 34 rows and 10 columns.
}
\usage{
china_province
}
\description{
province data for china
source: http://datav.aliyun.com/tools/atlas/

\describe{
  \item{name}{province name}
  \item{adcode}{adcode}
  ...
}
}
\keyword{datasets}
