% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletZH-package.R
\docType{package}
\name{leafletZH-package}
\alias{leafletZH}
\alias{leafletZH-package}
\title{leafletZH: Chinese Leaflet Map Relate Operation}
\description{
Provides 'sf' data for Chinese provinces and cities, methods for plotting shape maps of Chinese provinces and cities, Convert Coordinates Between Different Systems, and a layer for 'leaflet' with Gaode tiles. It is designed to facilitate geographical data visualization in China.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://damonsoul.github.io/leafletZH/}
}

}
\author{
\strong{Maintainer}:  Damonsoul \email{chenmaowei96@gmail.com}

}
\keyword{internal}
