% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_city}
\alias{china_city}
\title{city data for China}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 476 rows and 10 columns.
}
\source{
http://datav.aliyun.com/tools/atlas/

\describe{
  \item{name}{The name of the district }
  \item{adcode}{The administrative code for the district, a unique identifier (e.g., "110101")}
  \item{childrenNum}{The number of lower-level administrative divisions within the district (usually 0 for districts)}
  \item{level}{The administrative level of the area, which is "district" for all entries in this dataset}
  \item{subFeatureIndex}{An index representing the sub-features within the district}
  \item{centroid}{The geographical centroid of the district, represented as a string of coordinates }
  \item{center}{The center point of the district, also represented as a string of coordinates}
  \item{parent}{A JSON string representing the parent administrative entity, usually the province-level data}
  \item{acroutes}{A JSON array of administrative codes that represent the full administrative hierarchy leading to the district}
  \item{geometry}{Spatial geometry of the district, stored as an sf object in MULTIPOLYGON format}
}
}
\usage{
china_city
}
\description{
This dataset contains spatial (sf) data for city in China, with various attributes specific to each district.
}
\keyword{datasets}
