% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_province}
\alias{china_province}
\title{province data for China}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 34 rows and 10 columns.
}
\source{
http://datav.aliyun.com/tools/atlas/

\describe{
  \item{name}{The name of the province }
  \item{adcode}{The administrative code for the province, a unique identifier (e.g., "110000")}
  \item{childrenNum}{The number of administrative divisions (e.g., counties) within the province}
  \item{level}{The administrative level of the area, which is generally "province" for the entries in this dataset}
  \item{subFeatureIndex}{An index representing the sub-features within the province}
  \item{centroid}{The geographical centroid of the province, represented as a string of coordinates }
  \item{center}{The center point of the province, also represented as a string of coordinates}
  \item{parent}{A JSON string representing the parent administrative entity, usually the country-level data}
  \item{acroutes}{A JSON array of administrative codes that represent the administrative hierarchy leading to the province}
  \item{geometry}{Spatial geometry of the province, stored as an sf object in MULTIPOLYGON format}
}
}
\usage{
china_province
}
\description{
This dataset contains spatial (sf) data for provinces in China, including various attributes related to each province.
}
\keyword{datasets}
