% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletZH-package.R
\docType{package}
\name{leafletZH-package}
\alias{leafletZH}
\alias{leafletZH-package}
\title{leafletZH: Adds a Chinese Choropleth Leaflet Map}
\description{
Provides 'sf' data for Chinese provinces and cities, methods for plotting shape maps of Chinese provinces and cities, and a layer for 'leaflet' with Gaode tiles. It is designed to facilitate geographical data visualization in China.
}
\author{
\strong{Maintainer}:  Damonsoul \email{chenmaowei96@gmail.com}

}
\keyword{internal}
